Rename Provider Package
This commit renames the `provider` package to `cloudprovider` and
provides a more structured approach for each cloud provider.

Within the `aws` cloud provider, the ancillary functions have been
organized into separate files.

All commented out imports have also been removed.
